Move account construction to a module accessible from all tests.
