+++ concordium-consensus/tests/scheduler/SchedulerTests/Helpers.hs
@@ -9,7 +9,15 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module SchedulerTests.Helpers where
+module SchedulerTests.Helpers (
+    module SchedulerTests.Helpers,
+    DummyData.makeTestAccountFromSeed,
+    DummyData.keyPairFromSeed,
+    DummyData.accountAddressFromSeed,
+    DummyData.makeTestAccount,
+    DummyData.makeTestCredentialFromSeed,
+    DummyData.makeTestCredential,
+) where
 
 import qualified Control.Monad.Except as Except
 import Control.Monad.RWS.Strict
@@ -18,17 +26,12 @@ import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import Data.Word
 import Lens.Micro.Platform
-import qualified System.Random as Random
 import Test.HUnit
 
-import qualified Concordium.Crypto.DummyData as DummyData
 import qualified Concordium.Crypto.SHA256 as Hash
-import qualified Concordium.Crypto.SignatureScheme as SigScheme
-import qualified Concordium.ID.DummyData as DummyData
 import qualified Concordium.ID.Types as Types
 import qualified Concordium.Types.Accounts as Types
 import qualified Concordium.Types.Accounts.Releases as Types
-import qualified Concordium.Types.DummyData as DummyData
 import Concordium.Types.SeedState (initialSeedState)
 import qualified Concordium.Wasm as Wasm
 
@@ -74,70 +77,6 @@ simpleTransferCostWithMemo2 memoSize =
         (Types.transactionHeaderSize + 41 + 2 + memoSize)
         1
         + Cost.simpleTransferCost
-
--- |Generate an account credential with a single keypair and sufficiently late expiry date.
-makeTestCredential ::
-    SigScheme.VerifyKey ->
-    Types.AccountAddress ->
-    Types.AccountCredential
-makeTestCredential key accountAddress =
-    DummyData.dummyCredential
-        DummyData.dummyCryptographicParameters
-        accountAddress
-        key
-        DummyData.dummyMaxValidTo
-        DummyData.dummyCreatedAt
-
--- |Generate an account credential with a single keypair and sufficiently late expiry date
--- deterministically from a seed.
-makeTestCredentialFromSeed :: Int -> Types.AccountCredential
-makeTestCredentialFromSeed seed =
-    let keyPair = keyPairFromSeed seed
-        address = accountAddressFromSeed seed
-    in  makeTestCredential (SigScheme.correspondingVerifyKey keyPair) address
-
--- |Generate an account providing a verify key, account address and the initial balance.
--- The generated account have a single credential and single keypair, which has sufficiently late
--- expiry date.
-makeTestAccount ::
-    (Types.IsAccountVersion av, Blob.MonadBlobStore m) =>
-    SigScheme.VerifyKey ->
-    Types.AccountAddress ->
-    Types.Amount ->
-    m (BS.PersistentAccount av)
-makeTestAccount key accountAddress amount = do
-    let credential = makeTestCredential key accountAddress
-    account <- BS.newAccount DummyData.dummyCryptographicParameters accountAddress credential
-    BS.addAccountAmount amount account
-
--- |Generate a test account keypair deterministically from a seed.
--- Note this is also used internally by `makeTestAccountFromSeed` and providing the same seed
--- results in the same keypair as used for the generated account.
-keyPairFromSeed :: Int -> SigScheme.KeyPair
-keyPairFromSeed =
-    uncurry SigScheme.KeyPairEd25519
-        . fst
-        . DummyData.randomEd25519KeyPair
-        . Random.mkStdGen
-
--- |Generate an account address deterministically from a seed.
--- Note this is also used internally by `makeTestAccountFromSeed` and providing the same seed
--- results in the same account address as used for the generated account.
-accountAddressFromSeed :: Int -> Types.AccountAddress
-accountAddressFromSeed = DummyData.accountAddressFrom
-
--- |Generate a test account with the provided amount as balance. The generated account have a single
--- credential and single keypair, which has sufficiently late expiry date. The keypair and address
--- is generated deterministically from a seed.
-makeTestAccountFromSeed ::
-    (Types.IsAccountVersion av, Blob.MonadBlobStore m) =>
-    Types.Amount ->
-    Int ->
-    m (BS.PersistentAccount av)
-makeTestAccountFromSeed amount seed =
-    let keyPair = keyPairFromSeed seed
-        address = accountAddressFromSeed seed
-    in  makeTestAccount (SigScheme.correspondingVerifyKey keyPair) address amount
 
 -- | Monad that implements the necessary constraints to be used for running the scheduler.
 newtype PersistentBSM pv a = PersistentBSM
